requirements.txt
django
djangorestframework
mysqlclient
python-decouple
djangorestframework-simplejwt
opeanai